requirements.txt
uvicorn
pydantic
fastapi
pydantic-settings
python-dotenv
requests